feat: add to license
The license has been rewritten to be more explicit
about only applying to the written context of
posts. The translations have been set to en.yml
to keep that version as the source of truth as
I do not have verifiable translations for the
changes to the license for all locales.